+++ backend/test/fixtures/sample.ts
@@ -1,0 +1,109 @@
+export const sampleHtml = `
+<div style="border:1px solid #990000;padding-left:20px;margin:0 0 10px 0;">
+<h4>A PHP Error was encountered</h4>
+</div><!DOCTYPE html>
+<html xml:lang="kr" lang="kr" dir="ltr" xmlns="http://www.w3.org/1999/xhtml">
+<head>
+    <meta charset="UTF-8">
+    <title>평택당진도선사회</title>
+</head>
+<body id="top">
+<div id="mainwrap">
+    <div id="SubAllContentWrap">
+        <div class="Allwrap">
+            <div style="width:100%;">
+                <div class="tabdiv">
+                    <div id="tab_1" class="tab_view">
+                        <div class="portjewrap">
+                            <div>
+                                <div style="font-size:13px; font-family:'Noto Sans CJK KR Light'">
+                                    <div>
+                                        <table width="100%" border="0" cellpadding="0" cellspacing="0">
+                                            <tr>
+                                                <td><table cellpadding="3" cellspacing="1" border="0" width="100%" bgcolor="#c0c5cb" height="30">
+                                                        <tr>
+                                                            <td width="11%" height="25" align="center" bgcolor="#f2f2f2">날짜</td>
+                                                            <td height="25" bgcolor="#ffffff" width="30%">2025-11-29 토요일</td>
+                                                            <td height="25" bgcolor="#ffffff">NIGHT TIME : 17:30 ~ 07:00      일출: 07:24  일몰: 17:15</td>
+                                                        </tr>
+                                                    </table></td>
+                                            </tr>
+                                            <tr>
+                                                <td>
+                                                    <form name="listForm" id="listForm" method="post">
+                                                    <table cellpadding="3" cellspacing="1" border="0" width="100%" bgcolor="#c0c5cb" >
+                                                        <thead>
+                                                            <tr>
+                                                                <td width="30" height="30" align="center" bgcolor="#f2f2f2">번호</td>
+                                                                <td width="50" align="center" bgcolor="#f2f2f2">도선상태</td>
+                                                                <td width="50" align="center" bgcolor="#f2f2f2">도선사</td>
+                                                                <td width="70" align="center" bgcolor="#f2f2f2">일자</td>
+                                                                <td width="50" align="center" bgcolor="#f2f2f2">시간</td>
+                                                                <td width="50" align="center" bgcolor="#f2f2f2">긴특</td>
+                                                                <td align="center" bgcolor="#f2f2f2">선명</td>
+                                                                <td width="70" align="center" bgcolor="#f2f2f2" colspan="2">도선구간</td>
+                                                                <td width="50" align="center" bgcolor="#f2f2f2">접안</td>
+                                                                <td width="70" align="center" bgcolor="#f2f2f2">톤수</td>
+                                                                <td width="50" align="center" bgcolor="#f2f2f2">홀수</td>
+                                                                <td width="150" align="center" bgcolor="#f2f2f2">대리점</td>
+                                                                <td width="100" align="center" bgcolor="#f2f2f2">예선</td>
+                                                                <td width="50" align="center" bgcolor="#f2f2f2">강취</td>
+                                                                <td width="55" align="center" bgcolor="#f2f2f2">호출부호</td>
+                                                            </tr>
+                                                        </thead>
+                                                        <tbody>
+                                                            <tr style="background-color:#ffffff">
+                                                                <td style="width:36px;">1</td>
+                                                                <td style="width:66px;">완료</td>
+                                                                <td style="width:56px;">CS</td>
+                                                                <td style="width:46px;">11-29</td>
+                                                                <td style="width:46px;">00:00</td>
+                                                                <td style="width:46px;"></td>
+                                                                <td style="width:96px;">BBC ELISABETH</td>
+                                                                <td style="width:36px;"><span style="color: ">E12</span></td>
+                                                                <td style="width:36px;"><span style="color: red">IPA</span></td>
+                                                                <td style="width:42px;">좌현</td>
+                                                                <td style="width:48px;">18,189</td>
+                                                                <td style="width:40px;">7.6</td>
+                                                                <td style="width:130px;">세정해운</td>
+                                                                <td style="width:56px;">흥해</td>
+                                                                <td style="width:56px;">미래</td>
+                                                                <td style="width:46px;">V2HL2</td>
+                                                            </tr>
+                                                            <tr style="background-color:#E6E6E6">
+                                                                <td style="width:36px;">2</td>
+                                                                <td style="width:66px;">완료</td>
+                                                                <td style="width:56px;">MJ</td>
+                                                                <td style="width:46px;">11-29</td>
+                                                                <td style="width:46px;">00:00</td>
+                                                                <td style="width:46px;"></td>
+                                                                <td style="width:96px;">SHUO LONG 19</td>
+                                                                <td style="width:36px;"><span style="color: ">W7</span></td>
+                                                                <td style="width:36px;"><span style="color: red">IPA</span></td>
+                                                                <td style="width:42px;">우현</td>
+                                                                <td style="width:48px;">7,877</td>
+                                                                <td style="width:40px;">5.3</td>
+                                                                <td style="width:130px;">A-PLUS</td>
+                                                                <td style="width:56px;">대3</td>
+                                                                <td style="width:56px;">코마린</td>
+                                                                <td style="width:46px;">3E7618</td>
+                                                            </tr>
+                                                        </tbody>
+                                                    </table>
+                                                    </form>
+                                                </td>
+                                            </tr>
+                                        </table>
+                                    </div>
+                                </div>
+                            </div>
+                        </div>
+                    </div>
+                </div>
+            </div>
+        </div>
+    </div>
+</div>
+</body>
+</html>
+`;